Fixes and upgrades: double precision data values; cz_error_handler arguments; escape '+' in clinical data

+++ ddl/postgres/tm_cz/de_subj_protein_data_release.sql
@@ -4,7 +4,7 @@
 CREATE TABLE de_subj_protein_data_release (
     trial_name character varying(15),
     component character varying(15),
-    intensity bigint,
+    intensity double precision,
     patient_id bigint,
     subject_id character varying(100),
     gene_symbol character varying(100),
@@ -12,10 +12,10 @@ CREATE TABLE de_subj_protein_data_release (
     assay_id bigint,
     timepoint character varying(20),
     n_value bigint,
-    mean_intensity bigint,
-    stddev_intensity bigint,
-    median_intensity bigint,
-    zscore bigint,
+    mean_intensity double precision,
+    stddev_intensity double precision,
+    median_intensity double precision,
+    zscore double precision,
     release_study character varying(15)
 );
